Refactor the api files
